+++ CHANGELOG.md
@@ -64,9 +64,7 @@
 - [#2071](https://github.com/lapce/lapce/pull/2071): Add command and keybinds to delete line
 - [#2073](https://github.com/lapce/lapce/pull/2073): Add Ctrl+{a,e,k} keybinds on macOS
 - [#2128](https://github.com/lapce/lapce/pull/2128): Add Lapce app icon to logo collection
-- [#2127](https://github.com/lapce/lapce/pull/2127): Extended double-click options with file-only and file + folders mode
-- [#1944](https://github.com/lapce/lapce/pull/1944): Add filter input in git branch selection
-  - ![image](https://user-images.githubusercontent.com/4404609/211232461-293e3b31-4e17-457e-825c-3018699a6fc2.png)
+- [#2149](https://github.com/lapce/lapce/pull/2149) Support dynamic registration of Text Document Synchronization in LSP
 
 ### Bug Fixes
 - [#1911](https://github.com/lapce/lapce/pull/1911): Fix movement on selections with left/right arrow keys